Feat: Changed how the character control works.
The previous control system only changed the direction fo the character
and the actual movement was handled with an update method which was called by
the renderer. This implementation did not suit the future usage. The movement
is handled by the character and controller where the renderer only renders the
outcome.