Feat: Changed the how the map was implemented.
Old map was hard to pass around and had memory issues.
I created a new class for the map. And then passed it using a
shared pointer. This will ensure it will be there when needed.

Also, now it is easy to run checks on the map. Because there is
only one map shared accross all users, the methods to check the
existance of items will be implemented in the map class itself.

Using shared pointer creates a bit of an overhead, but I think
it is better than creating multiple copies and try to sync
them.